Tweaks on driver layout, tweaks on joystick and driving for cap lift.
