+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto100Red.java
@@ -33,6 +33,7 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 package org.firstinspires.ftc.teamcode;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 
 /**
  * This file illustrates the concept of driving a path based on encoder counts.
@@ -61,9 +62,9 @@ import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 
-@Autonomous(name="Auto Beta Red", group="DM")
-// @Disabled
-public class AutoShootCapRed extends AutoBetaBlue {
+@Autonomous(name="Auto Red Full 100", group="DM")
+//@Disabled
+public class Auto100Red extends Auto100Blue {
 
 
     public boolean amIBlue() {
