+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LineFinderTest.java
@@ -38,6 +38,7 @@ import android.view.View;
 import com.qualcomm.ftcrobotcontroller.R;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
 /*
@@ -68,7 +69,7 @@ import com.qualcomm.robotcore.hardware.DcMotor;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 @TeleOp(name = "Test Line Finder", group = "Sensor")
-// @Disabled                           // Comment this out to add to the opmode list
+@Disabled                           // Comment this out to add to the opmode list
 public class LineFinderTest extends LinearOpMode {
 
   HardwareDM robot = new HardwareDM();
