+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TestRGB.java
@@ -37,6 +37,7 @@ import android.view.View;
 
 import com.qualcomm.ftcrobotcontroller.R;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.ColorSensor;
 import com.qualcomm.robotcore.hardware.DeviceInterfaceModule;
@@ -70,7 +71,7 @@ import com.qualcomm.robotcore.hardware.DigitalChannelController;
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  */
 @TeleOp(name = "Test RGB Sensors", group = "Sensor")
-// @Disabled                           // Comment this out to add to the opmode list
+@Disabled                           // Comment this out to add to the opmode list
 public class TestRGB extends LinearOpMode {
 
   HardwareDM robot = new HardwareDM();
